Update styles for entire page
I honestly forgot to commit after wracking my brain for hours on doing the best practices in flexbox. I had fun playing around with everything and got the style down pretty close to what the model wants it to look like. The only things I did not add were different fonts and font sizes, and clickable links for some of the icons. I may decide to do that on a later date.